HEC-263: DisplayContract strips _id from reference column headers
- UILabelContract#label now strips trailing " Id" from names ending with _id
- DisplayContract gains reference_column_label(ref) and strip_id_suffix(name)
- Ruby UIGenerator index/show routes now include agg.references as columns/fields
- Go data_routes includes reference columns in index and show routes
- Go AggregateGenerator includes reference fields in struct definition
- Extracted UIGenerator::DataRoutes module to keep files under 200 lines
